+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Locked in gear".
 - Tests for "Robbery optimisation".
 - Tests for "The voucher".
+- Tests for "Connect the hyper-dots".
 
 ## [1.12.0] - 2022-09-01
 ### Added